BLUI-7251 change component name and prop updation
