+++ apps/showcase/src/components/brightlayer-ui/data-display/DataTable.tsx
@@ -1,0 +1,281 @@
+import React, { useState, useMemo, useCallback } from 'react';
+import { Box, Button, IconButton, Tooltip, Typography } from '@mui/material';
+import UndoIcon from '@mui/icons-material/Undo';
+import RedoIcon from '@mui/icons-material/Redo';
+import { DataTable, type DataTableColumnDef, type DataTableState } from '@brightlayer-ui/react-components';
+
+const componentContainerStyles = {
+    display: 'flex',
+    flexDirection: 'column',
+    flex: 1,
+    mb: 4,
+};
+
+type User = {
+    id: string;
+    firstName: string;
+    lastName: string;
+    email: string;
+    age: number;
+    state: string;
+    isActive: boolean;
+};
+
+const initialData: User[] = [
+    {
+        id: '1',
+        firstName: 'John',
+        lastName: 'Doe',
+        email: 'john.doe@example.com',
+        age: 30,
+        state: 'California',
+        isActive: true,
+    },
+    {
+        id: '2',
+        firstName: 'Jane',
+        lastName: 'Smith',
+        email: 'jane.smith@example.com',
+        age: 25,
+        state: 'New York',
+        isActive: false,
+    },
+    {
+        id: '3',
+        firstName: 'Bob',
+        lastName: 'Johnson',
+        email: 'bob.johnson@example.com',
+        age: 35,
+        state: 'Texas',
+        isActive: true,
+    },
+];
+
+const states = [
+    'California',
+    'New York',
+    'Texas',
+    'Florida',
+    'Illinois',
+    'Pennsylvania',
+    'Ohio',
+    'Georgia',
+    'North Carolina',
+    'Michigan',
+];
+
+export const DataTableExample: React.FC = () => {
+    const [users, setUsers] = useState<User[]>(initialData);
+    const [isLoading] = useState(false);
+    const [isSaving, setIsSaving] = useState(false);
+    const [tableState, setTableState] = useState<DataTableState | null>(null);
+
+    const columns = useMemo<Array<DataTableColumnDef<User>>>(
+        () => [
+            // {
+            //     accessorKey: 'id',
+            //     header: 'ID',
+            //     cellType: 'text',
+            //     enableEditing: false,
+            //     size: 80,
+            // },
+            {
+                accessorKey: 'firstName',
+                header: 'First Name',
+                cellType: 'text',
+            },
+            {
+                accessorKey: 'lastName',
+                header: 'Last Name',
+                cellType: 'text',
+            },
+            {
+                accessorKey: 'age',
+                header: 'Age',
+                cellType: 'number',
+                size: 88,
+            },
+            {
+                accessorKey: 'email',
+                header: 'Email',
+                // headerAlign: 'right',
+                cellType: 'text',
+            },
+            {
+                accessorKey: 'state',
+                header: 'State',
+                // headerAlign: 'left',
+                cellType: 'select',
+                editSelectOptions: states,
+            },
+            {
+                accessorKey: 'isActive',
+                header: 'Active',
+                cellType: 'binary',
+                size: 100,
+            },
+        ],
+        []
+    );
+
+    const validateUser = (user: User): Partial<Record<keyof User, string | undefined>> => {
+        const errors: Partial<Record<keyof User, string | undefined>> = {};
+
+        if (!user.firstName) {
+            errors.firstName = 'First Name is required';
+        }
+        if (!user.lastName) {
+            errors.lastName = 'Last Name is required';
+        }
+        if (!/^[\w-.]+@([\w-]+\.)+[\w-]{2,4}$/.exec(user.email)) {
+            errors.email = 'Invalid email format';
+        }
+        if (!user.age || user.age < 0 || user.age > 150) {
+            errors.age = 'Age must be between 0 and 150';
+        }
+        if (!user.state) {
+            errors.state = 'State is required';
+        }
+
+        return errors;
+    };
+
+    const handleCreate = async (newUser: User): Promise<void> => {
+        setIsSaving(true);
+        try {
+            await new Promise<void>((resolve) => {
+                setTimeout(resolve, 1000);
+            });
+
+            setUsers((prev) => [...prev, { ...newUser, id: (prev.length + 1).toString() }]);
+        } finally {
+            setIsSaving(false);
+        }
+    };
+
+    const handleUpdate = async (updatedUser: User): Promise<void> => {
+        setIsSaving(true);
+        try {
+            await new Promise<void>((resolve) => {
+                setTimeout(resolve, 1000);
+            });
+
+            setUsers(users.map((u) => (u.id === updatedUser.id ? updatedUser : u)));
+        } finally {
+            setIsSaving(false);
+        }
+    };
+
+    const handleDelete = async (id: string | number): Promise<void> => {
+        setIsSaving(true);
+        try {
+            await new Promise<void>((resolve) => {
+                setTimeout(resolve, 500);
+            });
+
+            setUsers(users.filter((u) => u.id !== id));
+        } finally {
+            setIsSaving(false);
+        }
+    };
+
+    const handleDuplicate = useCallback((duplicatedUser: User): void => {
+        setUsers((prev) => [...prev, { ...duplicatedUser, id: (prev.length + 1).toString() }]);
+    }, []);
+
+    const handleSave = useCallback(async (): Promise<void> => {
+        if (!tableState?.save) return;
+        setIsSaving(true);
+        try {
+            await tableState.save();
+            // eslint-disable-next-line no-console
+            console.log('Table saved. Updated data:', tableState.tableData);
+        } finally {
+            setIsSaving(false);
+        }
+    }, [tableState]);
+
+    return (
+        <Box sx={componentContainerStyles}>
+            {' '}
+            {/* Custom toolbar — driven by onStateChange */}
+            <Box sx={{ display: 'flex', gap: '0.5rem', alignItems: 'center', mb: 1 }}>
+                <Tooltip title="Undo (Ctrl+Z)">
+                    <span>
+                        <IconButton size="small" onClick={tableState?.undo} disabled={!tableState?.canUndo}>
+                            <UndoIcon fontSize="small" />
+                        </IconButton>
+                    </span>
+                </Tooltip>
+                <Tooltip title="Redo (Ctrl+Shift+Z)">
+                    <span>
+                        <IconButton size="small" onClick={tableState?.redo} disabled={!tableState?.canRedo}>
+                            <RedoIcon fontSize="small" />
+                        </IconButton>
+                    </span>
+                </Tooltip>
+                {tableState?.hasPendingChanges && (
+                    <>
+                        <Button
+                            variant="contained"
+                            color="primary"
+                            size="small"
+                            onClick={(): void => {
+                                void handleSave();
+                            }}
+                            disabled={isSaving || !!tableState?.hasValidationErrors}
+                            sx={{ textTransform: 'none' }}
+                        >
+                            Save Changes
+                        </Button>
+                        <Button
+                            variant="outlined"
+                            color="error"
+                            size="small"
+                            onClick={(): void => {
+                                tableState.reset();
+                            }}
+                            disabled={isSaving}
+                            sx={{ textTransform: 'none' }}
+                        >
+                            Reset
+                        </Button>
+                        <Typography variant="caption" color="text.secondary">
+                            You have unsaved changes
+                        </Typography>
+                    </>
+                )}
+            </Box>
+            <DataTable
+                columns={columns}
+                data={users}
+                onCreate={handleCreate}
+                onUpdate={handleUpdate}
+                onDelete={handleDelete}
+                onDuplicate={handleDuplicate}
+                onValidate={validateUser}
+                isLoading={isLoading}
+                isSaving={isSaving}
+                enableCreate={true}
+                editable={false}
+                enableDelete={true}
+                enableDuplicate={true}
+                enableColumnPinning={true}
+                enableRowActions={true}
+                enableCellActions={true}
+                enableClickToCopy="context-menu"
+                createDisplayMode="row"
+                editDisplayMode="cell"
+                createButtonText="Add New User"
+                enableUndoRedo={true}
+                onStateChange={setTableState}
+                // enableSorting={true}
+                // enableColumnFilters={true}
+                // enableColumnActions={true}
+                deleteConfirmMessage={(row: User) =>
+                    `Are you sure you want to delete ${row.firstName} ${row.lastName}?`
+                }
+            />
+        </Box>
+    );
+};
