+++ apps/showcase/src/components/brightlayer-ui/data-display/index.tsx
@@ -1,5 +1,5 @@
 export * from './ChannelValue';
-export * from './EditableTable';
+export * from './DataTable';
 export * from './EmptyState';
 export * from './Hero';
 export * from './InfoListItem';
