+++ docs/src/componentDocs/DataTable/examples/index.tsx
@@ -1,0 +1,4 @@
+export * from './BasicDataTable';
+export * from './RowEditingDataTable';
+export * from './CustomCellsDataTable';
+export * from './ExternalToolbarDataTable';
