+++ docs/src/componentDocs/DataTable/markdown/DataTableExamples.mdx
@@ -1,0 +1,26 @@
+import {
+    BasicDataTable,
+    RowEditingDataTable,
+    CustomCellsDataTable,
+    ExternalToolbarDataTable,
+} from '../examples';
+
+## Editable Table
+
+The `<DataTable>` component provides inline cell editing with built-in validation, CRUD operations, and an unsaved-change indicator (blue dot) per cell.
+
+<BasicDataTable />
+
+
+
+## Custom Cells and Cell Styling
+
+Use `Cell` to render any React element in a cell (e.g. a progress bar). Use `cellStyle` for conditional per-row MUI `sx` styles. Use `editVariant="select"` with `editSelectOptions` for dropdown editing.
+
+<CustomCellsDataTable />
+
+## External Toolbar with Save, Undo, Redo and Reset
+
+Use `enableUndoRedo` together with `onStateChange` to receive an `DataTableState` snapshot. Wire the stable callbacks (`undo`, `redo`, `save`, `reset`) and reactive flags (`canUndo`, `canRedo`, `hasPendingChanges`, `hasValidationErrors`) to your own toolbar controls placed anywhere in the page.
+
+<ExternalToolbarDataTable />
