--- docs/src/componentDocs/EditableTable/examples/index.tsx
@@ -1,4 +1,0 @@
-export * from './BasicEditableTable';
-export * from './RowEditingEditableTable';
-export * from './CustomCellsEditableTable';
-export * from './ExternalToolbarEditableTable';
