--- packages/component-library/src/core/EditableTable/index.ts
@@ -1,13 +1,0 @@
-export * from './EditableTable';
-export * from './types';
-export * from './cells';
-
-// Re-export commonly used types from material-react-table
-export type {
-    MRT_ColumnDef,
-    MRT_Row,
-    MRT_Cell,
-    MRT_Column,
-    MRT_TableOptions,
-    MRT_TableInstance,
-} from 'material-react-table';
