+++ packages/component-library/src/core/index.ts
@@ -6,7 +6,7 @@ export * from './InfoListItem';
 export * from './ListItemTag';
 export * from './Drawer';
 export * from './DrawerLayout';
-export * from './EditableTable';
+export * from './DataTable';
 export * from './ScoreCard';
 export * from './Utility';
 export * from './UserMenu';
